ai_generator/requirements.txt
# requirements.txt for the AI generator service

requests
Pillow
openai
boto3
